moved procces user input function to keyboard.c
