+++ kernel/kernel.c
@@ -8,7 +8,7 @@ void main()
     isr_install();
 
     //print project name and our name
-    char* osName = "TSOS\nBy - Shon and Tedy\n";
+    char* osName = "TSOS\nBy - Shon and Tedy\n> ";
     clear_screen();
 
     print(osName);
@@ -17,13 +17,4 @@ void main()
     init_keyboard();
 }
 
-void procces_user_input(const char* input)
-{
-    if (!strcmp(input, "END"))
-    {
-        print("turning the TsosOS off");
-        asm volatile("hlt");
-    }
-    print(input);
-    print("\n > ");
-}
+
