--- kernel/kernel.h
@@ -1,12 +1,0 @@
-#ifndef KERNEL_H
-#define KERNEL_H
-
-/*
- * procces the user input so if the input is END the os will shut down
- *
- * @param input The input of the user 
- */
-void procces_user_input(const char* input);
-
-
-#endif
